Fix bug in which emails would be fetched for initial indexing
